+++ src/dx-portal/next.config.js
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  output: "standalone",
+  reactStrictMode: true,
+};
+module.exports = nextConfig;